Estado de configuraciones se registra en sharedpreferences
