+++ app/src/main/java/com/tecmov2025/manoslocales/ActivityHome/MainNavigation.kt
@@ -4,10 +4,6 @@ import androidx.compose.runtime.Composable
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.tecmov2025.manoslocales.Database.AppDatabase
-import com.tecmov2025.manoslocales.Networking.ApiRepository
-import com.tecmov2025.manoslocales.Networking.RetrofitClient
-import com.tecmov2025.manoslocales.Utils.ExampleProductList
 import com.tecmov2025.manoslocales.Utils.ProductScreen
 import com.tecmov2025.manoslocales.Utils.ProductViewModel
 import com.tecmov2025.manoslocales.Utils.Screens
