+++ app/src/main/java/com/tecmov2025/manoslocales/Utils/Examples.kt
@@ -1,5 +1,5 @@
 package com.tecmov2025.manoslocales.Utils
-
+/**
 class ExampleProductList(
     val productosList: List<Producto> = List(20) { index ->
         when(index) {
@@ -303,3 +303,4 @@ class  ExamplesVendedoresList
 }
 
 
+*/
