+++ app/src/main/java/com/tecmov2025/manoslocales/Utils/Opciones.kt
@@ -1,0 +1,10 @@
+package com.tecmov2025.manoslocales.Utils
+
+import androidx.compose.ui.graphics.vector.ImageVector
+
+
+data class Opcion (
+    val text: String,
+    val onclick : ()-> Unit = {},
+    val icon: ImageVector? = null
+)
